+++ scala/AoC-2021/src/main/scala/days/Day3.scala
@@ -5,7 +5,7 @@ import scala.util.chaining.scalaUtilChainingOps
 
 class Day3 {
 
-  def solve(input: Seq[String]): (Int, Int) = {
+  def solve(input: Seq[String]): (Long, Long) = {
 
     val gammaRate =
       input
